Move GamificationToolbarReducer to presentation package
