+++ shared/src/commonMain/kotlin/org/hyperskill/app/gamification_toolbar/injection/GamificationToolbarComponent.kt
@@ -1,7 +1,7 @@
 package org.hyperskill.app.gamification_toolbar.injection
 
 import org.hyperskill.app.gamification_toolbar.presentation.GamificationToolbarActionDispatcher
-import org.hyperskill.app.gamification_toolbar.presentation.reducer.GamificationToolbarReducer
+import org.hyperskill.app.gamification_toolbar.presentation.GamificationToolbarReducer
 
 interface GamificationToolbarComponent {
     val gamificationToolbarReducer: GamificationToolbarReducer
